(PC-1326) extract options for distance select
