+++ src/components/pages/search/FilterByDistance.js
@@ -2,27 +2,9 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 
-// hope that pass culture is just still playing on earth
-const INFINITE_DISTANCE = 20000
-
-const options = [
-  {
-    label: "Moins d'1 km",
-    value: 1,
-  },
-  {
-    label: 'Moins de 10 km',
-    value: 10,
-  },
-  {
-    label: 'Moins de 50 km',
-    value: 50,
-  },
-  {
-    label: 'Toutes distances',
-    value: INFINITE_DISTANCE,
-  },
-]
+import options, {
+  INFINITE_DISTANCE,
+} from '../../../helpers/search/distanceOptions'
 
 class FilterByDistance extends PureComponent {
   onChange = e => {
