+++ src/helpers/search/distanceOptions.js
@@ -1,0 +1,23 @@
+export const INFINITE_DISTANCE = 20000
+// hope that pass culture is just still playing on earth
+
+const options = [
+  {
+    label: "Moins d'1 km",
+    value: 1,
+  },
+  {
+    label: 'Moins de 10 km',
+    value: 10,
+  },
+  {
+    label: 'Moins de 50 km',
+    value: 50,
+  },
+  {
+    label: 'Toutes distances',
+    value: INFINITE_DISTANCE,
+  },
+]
+
+export default options
